requirements.txt
textual>=0.70.0
rich>=14.0.0
chafa-py>=1.2.0
pytermgui>=7.7.4

llama-index-core
llama-index-readers-file
llama-index-llms-openai
llama-index-embeddings-openai
langgraph>=0.2.0
pydantic-ai>=1.0.0

tree-sitter>=0.25.0
GitPython>=3.1.43

marker-pdf>=1.0.3
crawl4ai>=0.3.0
httpx[http2]>=0.27.0
selectolax>=0.3.21

networkx>=3.3
pylatexenc>=2.10

lancedb>=0.14.0
